requirements.txt
pandas
sentence-transformers
umap-learn
hdbscan
matplotlib
jupyter
bertopic
top2vec
-e git+git@github.com:jchryssanthacopoulos/lda2vec.git#egg=lda2vec
